requirements.txt
rpi.gpio
requests